requirements.txt
flask==3.0.3
flask-sqlalchemy==3.1.1
flask-login==0.6.3
apscheduler==3.10.4
python-dotenv==1.0.1
openpyxl==3.1.5
stripe==11.4.0
pypdf==5.1.0
anthropic==0.40.0
asn1crypto==1.5.1
gunicorn==23.0.0
